/Top can now take in days as a fraction

+++ main.py
@@ -63,7 +63,7 @@ def new_goals(message):
 
 @bot.message_handler(commands=["top"])
 def manictime(message):
-	days= int(message.text.split('/top')[1])
+	days= float(message.text.split('/top')[1])
 	text = get_top_for_days(days)
 	bot.send_message(LAKSHAY_CID,text=text)
